move the root command to automation command
