+++ tools/generator/README.md
@@ -1,0 +1,3 @@
+# Generator
+
+This is a command line tool for generating new releases for `github.com/Azure/azure-sdk-for-go`.
