+++ tools/generator/cmd/automation/clean.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-package cmd
+package automation
 
 import (
 	"fmt"
