+++ tools/generator/cmd/automation/generation.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-package cmd
+package automation
 
 import (
 	"fmt"
@@ -12,7 +12,7 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/tools/generator/autorest"
 	"github.com/Azure/azure-sdk-for-go/tools/generator/autorest/model"
-	"github.com/Azure/azure-sdk-for-go/tools/generator/cmd/validate"
+	"github.com/Azure/azure-sdk-for-go/tools/generator/cmd/automation/validate"
 	"github.com/Azure/azure-sdk-for-go/tools/internal/exports"
 	"github.com/Azure/azure-sdk-for-go/tools/internal/utils"
 )
