+++ tools/generator/cmd/automation/validate/validation.go
